+++ dart/test/diagnostic_logger_test.dart
@@ -11,37 +11,39 @@ void main() {
     fixture = Fixture();
   });
 
-  test('$DiagnosticLogger do not log if debug is disabled', () {
-    fixture.options.debug = false;
+  group(DiagnosticLogger, () {
+    test('does not log if debug is disabled', () {
+      fixture.options.debug = false;
 
-    fixture.getSut().log(SentryLevel.error, 'foobar');
+      fixture.getSut().log(SentryLevel.error, 'foobar');
 
-    expect(fixture.loggedMessage, isNull);
-  });
+      expect(fixture.loggedMessage, isNull);
+    });
 
-  test('$DiagnosticLogger log if debug is enabled', () {
-    fixture.options.debug = true;
+    test('logs if debug is enabled', () {
+      fixture.options.debug = true;
 
-    fixture.getSut().log(SentryLevel.error, 'foobar');
+      fixture.getSut().log(SentryLevel.error, 'foobar');
 
-    expect(fixture.loggedMessage, 'foobar');
-  });
+      expect(fixture.loggedMessage, 'foobar');
+    });
 
-  test('$DiagnosticLogger do not log if level is too low', () {
-    fixture.options.debug = true;
-    fixture.options.diagnosticLevel = SentryLevel.error;
+    test('does not log if level is too low', () {
+      fixture.options.debug = true;
+      fixture.options.diagnosticLevel = SentryLevel.error;
 
-    fixture.getSut().log(SentryLevel.warning, 'foobar');
+      fixture.getSut().log(SentryLevel.warning, 'foobar');
 
-    expect(fixture.loggedMessage, isNull);
-  });
+      expect(fixture.loggedMessage, isNull);
+    });
 
-  test('$DiagnosticLogger always log fatal', () {
-    fixture.options.debug = false;
+    test('always logs fatal', () {
+      fixture.options.debug = false;
 
-    fixture.getSut().log(SentryLevel.fatal, 'foobar');
+      fixture.getSut().log(SentryLevel.fatal, 'foobar');
 
-    expect(fixture.loggedMessage, 'foobar');
+      expect(fixture.loggedMessage, 'foobar');
+    });
   });
 }