+++ dart/test/sentry_options_test.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: deprecated_member_use_from_same_package
+
 import 'package:http/http.dart';
 import 'package:sentry/sentry.dart';
 import 'package:sentry/src/noop_client.dart';
@@ -35,13 +37,30 @@ void main() {
 
   test('SentryLogger sets a diagnostic logger', () {
     final options = defaultTestOptions();
-    // ignore: deprecated_member_use_from_same_package
     expect(options.logger, noOpLogger);
-    // ignore: deprecated_member_use_from_same_package
     options.logger = dartLogger;
 
-    // ignore: deprecated_member_use_from_same_package
     expect(options.logger, isNot(noOpLogger));
+  });
+
+  test('setting debug correctly sets logger', () {
+    final options = defaultTestOptions();
+    expect(options.logger, noOpLogger);
+    expect(options.diagnosticLogger, isNull);
+    options.debug = true;
+    expect(options.logger, isNot(options.debugLogger));
+    expect(options.diagnosticLogger!.logger, options.debugLogger);
+    expect(options.logger, options.diagnosticLogger!.log);
+
+    options.debug = false;
+    expect(options.logger, isNot(noOpLogger));
+    expect(options.diagnosticLogger!.logger, noOpLogger);
+    expect(options.logger, options.diagnosticLogger!.log);
+
+    options.debug = true;
+    expect(options.logger, isNot(options.debugLogger));
+    expect(options.diagnosticLogger!.logger, options.debugLogger);
+    expect(options.logger, options.diagnosticLogger!.log);
   });
 
   test('tracesSampler is null by default', () {
@@ -112,7 +131,6 @@ void main() {
 
   test('when enableTracing is set to true tracing is considered enabled', () {
     final options = SentryOptions.empty();
-    // ignore: deprecated_member_use_from_same_package
     options.enableTracing = true;
 
     expect(options.isTracingEnabled(), true);
@@ -120,7 +138,6 @@ void main() {
 
   test('when enableTracing is set to false tracing is considered disabled', () {
     final options = SentryOptions.empty();
-    // ignore: deprecated_member_use_from_same_package
     options.enableTracing = false;
     options.tracesSampleRate = 1.0;
     options.tracesSampler = (_) {